+++ Certificate SDK Sample visionOS/CertificateSDKSampleVisionOSApp.swift
@@ -1,0 +1,18 @@
+//
+//  SPDX-License-Identifier: MIT
+//  https://github.com/jamf/CertificateSDK
+//
+//  Copyright 2024, Jamf
+//
+import SwiftUI
+
+@main
+struct CertificateSDKSampleVisionOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .frame(minWidth: 700, minHeight: 600)
+        }
+        .windowResizability(.contentMinSize)
+    }
+}