+++ Package.swift
@@ -1,20 +1,20 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.9
 import PackageDescription
 
 let package = Package(
     name: "CertificateSDK",
     platforms: [
-        .iOS(.v10)
+        .iOS(.v12),
+        .visionOS(.v1)
     ],
     products: [
         .library(
-            name: "CertificateSDK",
-            targets: ["CertificateSDK"])
+            name: "CertificateSDK", targets: ["CertificateSDK"]
+        )
     ],
     targets: [
         .binaryTarget(
-            name: "CertificateSDK",
-            path: "CertificateSDK.xcframework"
+            name: "CertificateSDK", path: "CertificateSDK.xcframework"
         )
     ]
 )